move `migrate-bare-utilities` to core
